requirements.txt
# Django and Core Dependencies
Django==5.2.4
djangorestframework==3.14.0
django-cors-headers==4.3.1

# PDF Processing
PyPDF2==3.0.1
pdf2image==1.17.0
reportlab==4.0.9
cairosvg==2.7.1

# Image Processing (required by pdf2image)
Pillow==10.2.0

# WSGI Server for Production
gunicorn==21.2.0

# Database (PostgreSQL - recommended for production)
psycopg2-binary==2.9.9

# Environment Variables Management
python-decouple==3.8

# Whitenoise for serving static files
whitenoise==6.6.0
